refactor for integration
